+++ client/examples/03_simple_block.rs
@@ -15,7 +15,6 @@ async fn main() -> Result<()> {
     let client = Client::builder()
         .with_node(&node_url)?
         .with_pow_worker_count(1)
-        .with_node_sync_disabled()
         .finish()?;
 
     let block = client.block().finish().await?;