+++ client/examples/block/04_block_tagged_data.rs
@@ -16,10 +16,7 @@ async fn main() -> Result<()> {
     });
 
     // Create a client with that node.
-    let client = Client::builder()
-        .with_node(&node_url)?
-        .with_node_sync_disabled()
-        .finish()?;
+    let client = Client::builder().with_node(&node_url)?.finish()?;
 
     // Create and send the block with tag and data.
     let block = client