+++ client/examples/node_api_core/21_get_milestone_by_index_raw.rs
@@ -17,10 +17,7 @@ async fn main() -> Result<()> {
     });
 
     // Create a client with that node.
-    let client = Client::builder()
-        .with_node(&node_url)?
-        .with_node_sync_disabled()
-        .finish()?;
+    let client = Client::builder().with_node(&node_url)?.finish()?;
 
     // Fetch the latest milestone index from the node.
     let info = client.get_info().await?;