+++ client/examples/node_api_indexer/04_get_foundry_outputs.rs
@@ -19,7 +19,6 @@ async fn main() -> Result<()> {
     let client = Client::builder()
         // The node needs to have the indexer plugin enabled.
         .with_node(&node_url)?
-        .with_node_sync_disabled()
         .finish()?;
 
     // Take the address from command line argument or use a default one.