+++ client/examples/send_all.rs
@@ -26,7 +26,6 @@ async fn main() -> Result<()> {
     // Create a client instance
     let client = Client::builder()
         .with_node(&node_url)? // Insert your node URL here
-        .with_node_sync_disabled()
         .finish()?;
 
     let secret_manager_1 = SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(