+++ client/src/client/builder.rs
@@ -248,10 +248,10 @@ impl ClientBuilder {
         self
     }
 
-    /// Disables the node syncing process.
+    /// Ignores the node health status.
     /// Every node will be considered healthy and ready to use.
-    pub fn with_node_sync_disabled(mut self) -> Self {
-        self.node_manager_builder = self.node_manager_builder.with_node_sync_disabled();
+    pub fn with_ignore_node_health(mut self) -> Self {
+        self.node_manager_builder = self.node_manager_builder.with_ignore_node_health();
         self
     }